+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 
 - Initial release (v0.0.1)
 - FIX: Added other colors (v0.0.2)
+- FIX: brightened up muted text (v0.0.3)
 
 ## [Unreleased]